Update gelatine-v1_1.md
Updated with a new attempt to use this recipe.  Using a technique from working with food gels (custards, etc ...) a blow torch can remove bubbles easily and consistently.   A chamber vacuum did not provide the desired results.  It caused the entire solution to explode as expelled air was caught in the polymer.